+++ docs-new/components/react/test-react-component.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export function TestReactComponent() {
+  return <div>Test Componentzzz</div>
+}