+++ docs/pages/_404/[...path].tsx
@@ -3,7 +3,7 @@
 export const Custom404 = () => <div></div>;
 
 export const getServerSideProps = () => {
-  return { redirect: { destination: "/", permanent: false } }; 
+  return { redirect: { destination: "/", permanent: false } };
 };
 
 export default Custom404;