+++ docs/pages/concepts/_meta.tsx
@@ -1,16 +1,11 @@
 import { SideNavTitleWithIcon } from "@/components/sidenav/sidenav-title-with-icon";
-import { FaReact } from "react-icons/fa";
-import { PiGraphDuotone } from "react-icons/pi";
-import { LuBrush } from "react-icons/lu";
 import { HiOutlineServerStack } from "react-icons/hi2";
+import { PiGraphDuotone } from "react-icons/pi";
 
 
 export default {
-  "generative-ui": {
-    title: <SideNavTitleWithIcon title="Generative UI" icon={FaReact} />,
-  },
   "self-hosting": {
-    title: <SideNavTitleWithIcon title="Self Hosting" icon={HiOutlineServerStack} />,
+    title: <SideNavTitleWithIcon title="Self-Hosting" icon={HiOutlineServerStack} />,
   },
   "different-llm-adapters": {
     title: <SideNavTitleWithIcon title="LLM Adapters" icon={HiOutlineServerStack} />,
@@ -18,8 +13,5 @@ export default {
   "agents": {
     title: <SideNavTitleWithIcon title="Agents (LangChain)" icon={PiGraphDuotone} />,
   },
-  "customize-look-and-feel": {
-    title: <SideNavTitleWithIcon title="Customize Look & Feel" icon={LuBrush} />,
-  }
 }