+++ examples/coagents-ai-researcher/agent/ai_researcher/summarize.py
@@ -3,12 +3,32 @@ The summarize node is responsible for summarizing the information.
 """
 
 import json
-from langchain_openai import ChatOpenAI
-from langchain_core.messages import SystemMessage, ToolMessage
+from langchain_core.messages import HumanMessage
 from langchain_core.runnables import RunnableConfig
 from copilotkit.langchain import copilotkit_customize_config
-
+from pydantic import BaseModel, Field
 from ai_researcher.state import AgentState
+from ai_researcher.model import get_model
+
+class Reference(BaseModel):
+    """Model for a reference"""
+
+    title: str = Field(description="The title of the reference.")
+    url: str = Field(description="The url of the reference.")
+
+class SummarizeTool(BaseModel):
+    """
+    Summarize the final result. Make sure that the summary is complete and 
+    includes all relevant information and reference links.
+    """
+
+    markdown: str = Field(
+        description="""
+        The markdown formatted summary of the final result. 
+        If you add any headings, make sure to start at the top level (#).
+        """
+    )
+    references: list[Reference] = Field(description="A list of references.")
 
 async def summarize_node(state: AgentState, config: RunnableConfig):
     """
@@ -20,7 +40,7 @@ async def summarize_node(state: AgentState, config: RunnableConfig):
         emit_intermediate_state=[
             {
                 "state_key": "answer",
-                "tool": "summarize",
+                "tool": "SummarizeTool",
             }
         ]
     )
@@ -32,57 +52,15 @@ These are all of the steps: {json.dumps(state["steps"])}
 Please summarize the final result and include all relevant information and reference links.
 """
 
-    summarize_tool = {
-        'name': 'summarize',
-        'description': """
-Summarize the final result. Make sure that the summary is complete and includes all relevant information and reference links.
-""",
-        'parameters': {
-            'type': 'object',
-            'properties': {
-                'markdown': {
-                    'description': 'The markdown formatted summary of the final result. If you add any headings, make sure to start at the top level (#)',
-                    'type': 'string'
-                },
-                'references': {
-                    'description': """A list of references.""",
-                    'type': 'array',
-                    'items': {
-                        'type': 'object',
-                        'properties': {
-                            
-                            'title': {
-                                'description': 'The title of the reference.',
-                                'type': 'string'
-                            },
-                            'url': {
-                                'description': 'The url of the reference.',
-                                'type': 'string'
-                            },
-                        },
-                        'required': ['title', 'url']
-                    }
-                }
-            },
-            'required': ['markdown', 'references']
-        }
-    }
-
-    response = await ChatOpenAI(model="gpt-4o").bind_tools([summarize_tool], parallel_tool_calls=False, tool_choice="summarize").ainvoke([
-        *state["messages"],
-        SystemMessage(
+    response = await get_model().bind_tools(
+        [SummarizeTool],
+        tool_choice="SummarizeTool"
+    ).ainvoke([
+        HumanMessage(
             content=system_message
-        )
+        ),
     ], config)
 
     return {
-        "messages": [           
-            response,
-            ToolMessage(
-                name=response.tool_calls[0]["name"],
-                content="summarized.",
-                tool_call_id=response.tool_calls[0]["id"]
-            )
-        ],
         "answer": response.tool_calls[0]["args"],
     }